[untangleworms]: Move functions that don't really depend on `self` `get_angles` `path_to_pixel_coords` `calculate_angle_shape_cost` `sample_control_points` `calculate_cumulative_lengths` `get_all_paths_recur` `build_incidence_lists` `calculate_path_length` `get_all_paths` `get_longest_path_coords` `make_incidence_matrix` `trace_segments` `get_graph_from_branching_areas_and_segments` `get_graph_from_binary`
